# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CBS Swagger backend server.",
        "title": "CBS Backend Swagger server",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/logs/getLatest/{secureId}": {
            "get": {
                "description": "Gets latest logs of secureId from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logging"
                ],
                "summary": "Gets latest logs of secureId",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Secure ID",
                        "name": "secureId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Log"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    }
                }
            }
        },
        "/logs/syncLogs": {
            "post": {
                "description": "Syncs logs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logging"
                ],
                "summary": "Syncs logs",
                "parameters": [
                    {
                        "description": "Log",
                        "name": "logs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Log"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    }
                }
            }
        },
...
        "/trackers": {
            "post": {
                "description": "Sync Trackers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracker"
                ],
                "summary": "Sync Trackers",
                "parameters": [
                    {
                        "description": "Tracker",
                        "name": "tracker",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Tracker"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Tracker"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    }
                }
            }
        },
        "/trackers/GetLatestTrackersPeriods/{secureId}": {
            "get": {
                "description": "Get latest tracker periods from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracker"
                ],
                "summary": "Get latest tracker periods",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Secure ID",
                        "name": "secureId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.DeltaDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    }
                }
            }
        },
        "/trackers/getLatest/{secureId}": {
            "get": {
                "description": "Get latest trackers from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracker"
                ],
                "summary": "Get latest trackers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Secure ID",
                        "name": "secureId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tracker"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorLog"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.DeltaDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "integer"
                },
                "delta": {
                    "type": "integer"
                }
            }
        },
...
        "models.Device": {
            "type": "object",
            "properties": {
                "androidId": {
                    "type": "string"
                },
                "brand": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "device": {
                    "type": "string"
                },
                "deviceModel": {
                    "type": "string"
                },
                "errorLogs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ErrorLog"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Log"
                    }
                },
                "product": {
                    "type": "string"
                },
                "sdk": {
                    "type": "string"
                },
                "secureId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userId": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.ErrorLog": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "deviceID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Location": {
            "type": "object",
            "properties": {
                "accuracy": {
                    "type": "number"
                },
                "altitude": {
                    "type": "number"
                },
                "bearing": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "provider": {
                    "type": "string"
                },
                "sensorType": {
                    "type": "string"
                },
                "speed": {
                    "type": "number"
                },
                "trackerID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Log": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dateTime": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "device": {
                    "$ref": "#/definitions/models.Device"
                },
                "deviceID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Tracker": {
            "type": "object",
            "properties": {
                "batteryLevel": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "string"
                },
                "device": {
                    "$ref": "#/definitions/models.Device"
                },
                "deviceID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Location"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuthorization": {
            "type": "basic"
        }
    }
}